Change to ask for an input
Also changed the number of the pins due to the pin number of the board being specified earlier in the program and I had used the bcim numbers. Testing the time.sleep() as well. Need to find a way to execute the program based on what key has been pressed can keeping virgil moving while that button is being pressed.